lab7_u1k2m/src/main/model/question/checker/ArithmeticAnswerChecker.java
package model.question.checker;

public class ArithmeticAnswerChecker extends AnswerChecker {

    private int answer;

    public ArithmeticAnswerChecker(int answer) {
        this.answer = answer;
    }

    @Override
    public boolean checkAnswer(String userResponse) {
        int userResponseAsInt;
        try {
            userResponseAsInt =  Integer.parseInt(userResponse);
        } catch (NumberFormatException e) {
            return false;
        }
        return Integer.parseInt(userResponse) == answer;
    }
}
